more code fixes

+++ yggdrasil-node/src/main/java/org/yggdrasil/node/network/messages/payloads/BlockHeaderPayload.java
@@ -104,8 +104,8 @@ public class BlockHeaderPayload implements MessagePayload {
             return new Builder();
         }
 
-        public Builder setVersion(BigInteger version) {
-            this.version = version.intValue();
+        public Builder setVersion(int version) {
+            this.version = version;
             return this;
         }
